fix(ci): resolve 69 mypy errors in new agent/core/strategy files
Fix all type errors introduced by the feature branch in 21 files:
- Add Callable import and replace builtin callable in type annotations
- Add missing return type annotations (-> None) to functions
- Fix async method overrides to match parent class signatures
- Fix overloaded function key arguments in max() calls
- Fix incompatible type assignments (float/bool, int/bool)
- Fix Decimal cast for tuple return types
- Add None guards for optional attributes
- Remove unused type: ignore comments
- Apply black formatting

+++ bot/strategies/hybrid/recovery_coordinator.py
@@ -455,7 +455,7 @@ class RecoveryCoordinator:
         if total_size == 0:
             return Decimal("0"), Decimal("0")
 
-        return total_cost / total_size, total_size
+        return Decimal(total_cost / total_size), Decimal(total_size)
 
     def _generate_dca_signals(
         self, state: RecoveryState, current_price: Decimal
